Add protype device registry with new shell function
The probe and scan commands now register new devices they find in a
registry.  This registry is currently lost on power loss and reboot.

The shell command 'mbus status' can be used to query status of the line
and also the M-Bus devices that have been found.

The new shell command 'mbus reset' resets both serial line settings and
the device registry.

Both the probe and scan functions have been extended to do an extra
M-Bus request for each located device to extract the primary and
secondary address, respectively, from the reply frame.

Note, the shell command 'mbus address' now updates the primary address
in the registry on successful operation.  See 'mbus status' to verify
the restult of the operation.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>